moving testing out of functions crate due to circular dependencies
