+++ datafusion/functions/Cargo.toml
@@ -91,8 +91,6 @@ uuid = { version = "1.7", features = ["v4"], optional = true }
 [dev-dependencies]
 arrow = { workspace = true, features = ["test_utils"] }
 criterion = "0.5"
-datafusion-optimizer = { workspace = true, default-features = true }
-datafusion-physical-expr = { workspace = true, default-features = true }
 rand = { workspace = true }
 tokio = { workspace = true, features = ["macros", "rt", "sync"] }
