+++ datafusion/functions/src/utils.rs
@@ -124,13 +124,6 @@ where
 pub mod test {
     use super::*;
     use arrow::datatypes::DataType;
-    use datafusion_common::DFSchema;
-    use datafusion_expr::execution_props::ExecutionProps;
-    use datafusion_expr::simplify::SimplifyContext;
-    use datafusion_expr::Expr;
-    use datafusion_optimizer::simplify_expressions::ExprSimplifier;
-    use datafusion_physical_expr::PhysicalExpr;
-    use std::sync::Arc;
 
     #[test]
     fn string_to_int_type() {
@@ -203,26 +196,6 @@ pub mod test {
                 }
             };
         };
-    }
-
-    /// Create a [`PhysicalExpr`] from an [`Expr`] after applying type coercion.
-    ///
-    /// Note: Can't use SessionContext::new().create_physical_expr(…) due to circular dependencies.
-    pub(crate) fn create_physical_expr_with_type_coercion(
-        expr: Expr,
-        df_schema: &DFSchema,
-    ) -> Result<Arc<dyn PhysicalExpr>> {
-        let props = ExecutionProps::default();
-        let context =
-            SimplifyContext::new(&props).with_schema(Arc::new(df_schema.clone()));
-        let simplifier = ExprSimplifier::new(context);
-        let coerced_expr = simplifier.coerce(expr, df_schema)?;
-        let physical_expr = datafusion_physical_expr::create_physical_expr(
-            &coerced_expr,
-            df_schema,
-            &props,
-        )?;
-        Ok(physical_expr)
     }
 
     pub(crate) use test_function;
