+++ datafusion/optimizer/Cargo.toml
@@ -54,6 +54,7 @@ regex-syntax = "0.8.0"
 [dev-dependencies]
 async-trait = { workspace = true }
 ctor = { workspace = true }
+datafusion-functions = { workspace = true }
 datafusion-functions-aggregate = { workspace = true }
 datafusion-functions-window = { workspace = true }
 datafusion-functions-window-common = { workspace = true }
